update design to buy creature in shop
